Update 07/07/2021 - 08/07/2021 - Converting to touch controls
Since this project needs to be accessible to a more casual audience, it needs to be playable on devices that everyone owns. To this end, I have been instructed to make this game work on smartphones. Touch controls are a difficult thing I’m not particularly used to, so my first step was to copy a touch-based joystick script that I made in an old TAFE class. This would serve adequately for standard two-axis movement and camera controls, but since the game is set in space, a zero-gravity environment with no up or down, I needed to account for my other controls – vertical movement, and a function to rotate the camera on the Z axis so the player is not awkwardly upside down compared to something they want to look like.
My original plan was to just have buttons for ascent, descent, clockwise and counter-clockwise rotation, like how the keyboard and mouse controls currently work, but I could not get Unity’s default Button class to show data for anything other than a regular click, when I needed separate events for pressing, holding and releasing the button. After some googling, I found out I had to make my own interactable object, with support for those things. After some experimenting and curiosity, I got distracted and did not make regular buttons, but I have made something even more interesting – a touch zone that supports dragging. As well as conventional presses, this interface element should allow for swipes and dragging. While overkill for what I am making, this functionality should be incredibly useful for a wide variety of mobile control inputs, and I can use this functionality to have analog input, allowing for finer control over vertical movement and camera rotation.
This function is still a work in progress, though. Currently, it uses screen space for all its output variables and distances, which is hard to use in a way that scales with different screens and aspect ratios. However, I am not sure what specifically to replace it with and how to do it, as it will require a lot of extra code to convert screen space to canvas or viewport space. But whatever system I use, this is a very valuable development in my skill with mobile development.