Update 13/07/2021 - 14/07/2021: Continued work on touch controls, positive progress on trackpad, drag zone and virtual joystick
PERFORMED MORE WORK ON TOUCH CONTROLS
Almost completed VirtualAnalogStick. I need a way to convert the centre of the RectTransform into a grid coordinate similar to how I converted the screen position input values into canvas space, so I can get the appropriate relative position of the player’s finger position. Otherwise, it seems to be working well. In better news, I have also seemingly completed DragZoneAsAnalogStick, although it needs more testing.
FIXING NON-RESETTING DELTA VALUES IN DRAG ZONE
Seemingly completed DragZoneAsTrackpad. Needs more testing but I solved the problems of the deltas not updating, with a clunky looking but effective system involving an IEnumerator.
For some inexplicable reason, the delta data in the PointerEventData from OnDrag will not reset to zero when the player stops moving their finger. This has the annoying effect of the game acting as if the player is still dragging their finger, even if they have stopped moving it. I thought I could solve this with IEndDragHandler and OnDragEnd, but this inexplicably does not activate until the player lifts their finger, making it functionally identical to OnPointerUp. I needed a way to reset the delta input to zero when the player stops moving their finger across the screen, but is still pressing it down.
In the DragZone class, I added a checking bool. When the player first initiates OnDrag, the function creates an IEnumerator, unimaginatively named CheckPlayerHasStoppedDragging(). This function has a while loop and a WaitForEndOfFrame() yield within it. Every frame, OnDrag will set the checking bool to true, and the IEnumerator will set it to false. When the player stops dragging their finger, OnDrag will stop running, but the IEnumerator will persist. It checks the state of the checking bool, and if it has not changed since the previous frame, this means OnDrag has stopped running. The IEnumerator will then reset the delta value and end itself, so that when the player drags their finger over the screen the process will start again.