Update 15/07/2021 - Fully functioning virtual analog sticks, and a first revision of my touch control scheme
Today was a good day.
I fixed the analog stick input positioning problems. Since the most accurate and reliable position value for RectTransforms, anchoredPosition, only shows the canvas position relative to its anchor point, I created some algorithms that use the joystick’s RectTransform anchor points to calculate the actual coordinates on the screen, ignoring its anchor point. This allowed me to accurately gauge the pivot point, and therefore the centre of the joystick’s rectangle, so I can judge the player’s relative finger position accordingly.
This caused some problems when I found that if the anchor point was in certain positions relative to the input zone, it would cause the positions to continue messing up. I realised this was because I was calculating the ‘true’ position in Awake(). The canvas seems to undergo some calculations of its own to make the dimensions accurate, before the game starts but after Awake(), so my code was working off of incorrect calculations. As a slightly less performant but simple solution, I changed the timing of said code so it activates not before the game starts, but whenever the player presses their finger down. While it does redundantly activate multiple times, this problem is now completely solved, and I have more important problems to solve. Another minor issue that I encountered is that when activating the bools that invert the final value, the handle position would also invert to match it, creating a bizarre effect where the handle would move to the opposite direction the player pulled in, despite ostensibly being the thing grabbed by the player. I had to invert the values back after saving the input value, before animating the handle. There’s probably a more efficient way to accomplish this, but what’s important is that my virtual analog stick seems to be complete!
I finally assembled a functioning first revision of a touch control scheme for the gameplay demo. It currently consists of 2 regular analog sticks and 2 single-axis sticks, therefore allowing control over the six axes of movement in 3D space – 3 for movement and 3 for rotation. The current setup is somewhat awkward to use, and I have not tested it on a real phone yet (just my laptop’s touchscreen), but from a programming perspective it appears to be completely bug-free.
Additionally, I created some new art assets for the single axis joystick, to match the existing assets I was using for more conventional rounded two-axis joysticks.
Notable touchscreen functions I need to work on:
•	Refining drag zone functionality
•	Adding an option for a dynamic background and handle to the DragZoneAsAnalogStick class
•	Adding an ability to eliminate analog input and use the VirtualAnalogStick like a D-pad (or just create a new VirtualDirectionalPad class)
•	Adding buttons that can sense being pressed down, held and released, as opposed to just clicked.
Adding the option to press down two fingers onto a DragZone, and twirl them to rotate something.