Update 19/07/2021 – Functioning pause menu and main menu, collision fix
I completed the new revision of the game’s touchscreen-based pause menu, which was actually relatively simple as I was able to adapt most of the code from my first-person shooter. I did add one new feature. I theorise that this game may eventually include multiplayer, for a more healthy social aspect. A pause menu in a traditional singleplayer game will obviously stop time so things don’t happen while the player is taking a break, but in a multiplayer game it wouldn’t be particularly fun for everybody to be paused just because one player pressed the pause button. I devised a small solution, involving a new class named GlobalPauseManager. A single one of this class exists in the scene, and every time a player presses the pause button, a function is activated which checks if the other players are all paused. It then stops time if everyone in the game is paused at once. If a player unpauses, time will resume. It might be unnecessary to have this code be in a separate class, when I could probably implement it as part of the regular pause function, so this is something I will need to rectify later.

I also created a functioning main menu. This was relatively simple, again because I was able to copy code over from my first-person shooter.

I also solved some basic annoying bugs with the player. The player would previously phase right through planets instead of colliding with them, and faraway planets would blink awkwardly in and out of existence due to being culled. The first problem was due to the MeshCollider class not existing with default settings, and the Convex setting was set to false. I simply added a line of code to set that to true when the planet is created. To fix the blinking problem, I simply increased the culling distance for the player’s camera.