Update 19/07/2021 - Streamlined and fixed time alteration related to pausing
I realised the GlobalPauseManager was completely redundant. The functions in it, I could add directly into the game's pause and resume functions, without sacrificing anything. I then deleted the GlobalPauseManager class and all references to it.

I also fixed a bug, where time would stay paused when transitioning to the main menu from the pause menu. This would cause the animations on the main menu to not play properly. I added another function to the MenuHandler class, ResetTime(), which I was able to add as an event for the quit to menu button. This acts as an easy way to unpause time when going to the main menu.