Update 09/08/2021 - Started adding support for both touch and keyboard + mouse controls
As per some instructions from the person I’m working on this for, I’m modifying the game’s controls so they work well with either touch or keyboard and mouse controls. At first it was just the latter, which I then replaced with the former. I also had to create some extra code to show/disable the touch controls and cursor when the game resumes, depending on which option the player picks. I set it to happen when the game resumes, rather than just on awaking, to account for the fact that the player may want to change their controls in an options menu while the game is paused. That being said, in the editor, it is possible to switch between these two control options by toggling a bool, but the player presently has no way to access this option. I started work on an options menu to allow this to happen, but this is not fully connected up yet.

I created an experimental function that takes a movement input from a virtual joystick and snaps it to one of a specified number of directions. This would be helpful for something like a virtual D-pad, where you want digital directional inputs but with the feel of a moving a single control rather than the awkward practice of simply putting virtual buttons on the screen. Of course, the function is specified with an integer for the number of possible directions, allowing a wide level of input options.

I changed a public variable in DragZone, DragDirectionFromOrigin, to act as a property instead of a method, to make it easier to reference in other classes.