Update 26/08/2021: Started attempting to implement gyro controls
I added some simple code that alters the player's camera movement based on the gyroscope input of the player's device, as well as a bool to turn it on and off. This feature is untested and undoubtedly needs more tweaking, but this is something I have been instructed to add.