Update 23/09/2021: File sorting, experimenting with options menus, land-based first-person controller
Sorted many of the files in the asset folder for a less cluttered navigation experience
Experimented with creating button prompts for options to enable/disable touch and gyro controls.
I experimented with creating a first-person controller for use in a level with gravity and a floor, using my custom touch controls. I did some experimenting but didn’t get very far with it. Since I put a lot of work into many of my control concepts, but realised that the setup I have currently only utilises some of it, I realised I needed to make something to show off my other assets.