Update 24/09/2021: Fully implemented and tested gyro controls, work on options menu
I created proper functioning gyroscope controls to rotate the player’s camera! Hooray! The controls are enabled and disabled via an easy-to-access toggle switch on the screen during normal gameplay, next to the regular camera movement controls. I implemented this so players in awkward positions could easily turn off the gyro controls to manually get their bearings and play in a comfortable position, something that can also be updated by shifting the analog stick.

I fully implemented the ability to smoothly transition between different menus, allowing for fully functional complex menu hierarchies. This was something I have struggled with in many projects, but I have finally figured out a solution. It turned out more complicated than I wanted, but it works well, involving updating Canvas Group values to show and hide certain window canvases without disabling child windows.

Started work on a functioning options menu. This is important since the client specified that the game have the option for multiple input types – keyboard and mouse on a computer, or touch controls on a smartphone. I started creating special sliders designed for adjusting camera sensitivity values, consisting of a slider and a secondary toggle for inverting the controls. I added all the necessary controls into the options menu, in a relatively easy to read format, and started work on a script to manage the menu, such as applying changes and updating the menu to reflect current settings. This has not been fully implemented, however, and as a result none of the options presently do anything. But it’s good progress.