Tweaked boosting mechanics in zero gravity first-person controller
Tweaked current movement speed variable to be a property that obtains the base movement speed and multiplies it by the boost multiplier, if the isBoosting bool is true, rather than having a value that is toggled whenever a function is run. This method will be more reliable and less prone to being stuck in a particular state contrary to what the current state is supposed to be. The touch control for toggling boosting simply changes the isBoosting bool depending on the current value.

The touch control for boosting also doesn't disappear when keyboard and mouse controls are active, since it is still updated when the player presses the keyboard button to boost, and acts as a handy HUD indicator.