Zero gravity control options menu now functioning
Previously, this interface existed but did not function, due to there being no apply or revert buttons. Now, it does apply settings properly.
When the options menu screen is opened, a refresh function will run to update all settings to match the player’s current settings. The apply and revert buttons are disabled.
When one of the stats is changed from the default, the apply and revert buttons are enabled using a listener. If the apply button is pressed, the player’s options will be updated to match the values in the menu. If the revert button is pressed, it will simply run the refresh button to change the options back to the player’s original stats, without any of the player’s values being changed.
I had to fix several bugs. For example, in Unity’s execution order, Awake() runs before OnEnable(), which runs before Start(). I want the menu’s options to refresh using the RefreshMenu() function to match the player’s current settings whenever it is opened, so I have a reference to it in OnEnable(). But that means it will run before the player variable is initialised (which I had in Awake()), causing null errors. If I ran the check directly inside RefreshMenu(), that would instead cause different null errors where my sensitivity sliders don’t have their controls initialised, since Awake() doesn’t seem to run and initialise them until their objects are enabled. To fix all of this, I had to come up with the roundabout solution of running the player initialisation code and RefreshMenu() inside Start(), running after everything else, then in OnEnable() have a check to make sure the player is not null before refreshing (so when it is first opened it does not run without Start() being run, but will run normally on all subsequent attempts.
Additionally, I made a rookie mistake of assigning some changed values to the wrong player stats, which was easily fixed.
I also changed the menu prefab name from 'Options' to Zero Gravity Control Options, since the options are solely for player controls.

In conclusion, this system works very smoothly. It isn’t completed yet – I need to add functions for saving options data between play sessions, resetting to default options, and making the current slider value clearer and easier to implement. But I’m very happy with how this has turned out.