Third-person spaceship controller
I created the SpaceshipMovement script, an arcade but fun and responsive movement controller that can simulate a spaceship. The ship controls fairly simply: Space/Shift for acceleration and reversing, W/S to control pitch, A/D for yaw and Q/E for roll. A ‘warp drive’ function is also present that disables control temporarily, turns and rapidly zooms towards an object in the scene, but this is not fully implemented yet and there is no way to select an area to warp to. Touch controls are also present for these purposes, but mouse control is not presently available. I also made a new scene to test this code in.
To make it compatible with the existing systems, I had to slightly modify the way the other movement controller, player handler and game state handler worked. I made a new class named MovementController, and changed both SpaceshipMovement and FirstPersonZeroGravityController to inherit from it. I made the SetControlsToComputerOrMobile() function in the FirstPersonZeroGravityController class instead be a virtual function in MovementController, so it can be overridden in SpaceshipMovement to account for the different touch controls available. This switch to inheritance was intended so the PlayerHandler and GameStateHandler classes would function properly with the different kinds of movement controllers for different gameplay genres present in this project.