Created base ‘OptionsMenu’ class and reworked ‘ZeroGravityControlOptions’ class to derive from it.
I did this for ease of creation when it comes to creating other types of options menus. The base OptionsMenu class is abstract and contains the code for the apply and revert buttons, and three abstract functions to alter in inherited classes.
The first, SetupOptions(), exists to prepare each option interactable in the menu by adding functions as listeners, primarily the OnOptionsChanged() function that enables the apply and revert buttons. This function exists as an easy way for the player to tell if an option has been changed. This function runs in Awake(), but is included as a separate function from Awake() so a developer cannot accidentally override essential functions running in Awake().
The second, ObtainCurrentValues(), will contain code to establish a reference to whatever objects have their stats altered, and update each option interactable with the appropriate current stat. This is run in a private Refresh() function, that also sets the apply and revert buttons to false to show that these values are unchanged from when they were last set. The Refresh() function will run in OnEnable, or when settings are applied as a sanity check and to update the apply and revert buttons.
The third, ApplySettings(), will make changes to the stats of the currently referenced object. This runs alongside Refresh() in a function called Apply(), to ensure that the two functions always run together, since Apply() is not virtual or abstract and therefore cannot be overridden.
The object reference and the interactables are established as fields in inherited classes, as each kind of options menu will have different ones to alter.

Since this class now exists, I had to remove all of this functionality from my ZeroGravityControlOptions class, then change it to derive from OptionsMenu instead of MonoBehaviour. I then moved its functions for setting up interactables, obtaining and applying values, and placed them appropriately in the SetupOptions(), ObtainCurrentValues() and ApplySettings() functions.