Static functions to convert sensitivity values to and from a slider range
I created two functions, SensitivityToSliderValue() and SliderValueToSensitivity(). The first is a void function that takes a current sensitivity value, a max sensitivity value, and a Slider object, and updates the Slider component to display an equivalent value, but within the slider’s range. The SliderValueToSensitivity() function is a float that converts it back to a value usable by the game process. Sometimes a sensitivity value can be very large or small and look weird to the end user, so these functions exist to add a layer of abstraction and, to the player in the options menu, present a more sensible appearing number range to pick from.