SpaceshipControlOptions class
I created a new class for updating the player’s spaceship controls. This menu is very similar to the ZeroGravityControlOptions menu, as they both derive from OptionsMenu, so there’s not really much to mention here. The options are mainly for steering, consisting of mouse and gyro sensitivity, and bools for inverting the pitch, yaw and roll.