Tweaked spaceship camera code to prevent it from lagging too far behind.
The code involves obtaining the position of the camera relative to the player transform, clamping its magnitude to the distance between the desired camera orientation and the player transform (i.e. the ideal distance), and adding the player's transform position to the resulting Vector3 to convert it back into a world position. Although there are some awkward looking visual effects when using the warp function, this system does a fairly decent job of ensuring the player's camera does not lag too far behind.