Fixed a bug where the options menu apply and revert settings wouldn't be disabled when new settings were applied.
The OptionsMenu class has two functions - ApplySettings(), an abstract function where specific settings are applied, and Apply(), a private class that runs ApplySettings() as well as Refresh() to update all the options. I accidentally assigned ApplySettings() instead of Apply() to the apply button, so the additional refreshing code was not running.