Updated spaceship steering controls to account for external physics rotation
Previously, the player spaceship would be unaffected by external rotation, but this felt awkward when the player crashed into something. But if I simply allowed external forces to affect the player's rotation, the player could be stuck endlessly spinning without being able to correct it because the steering controls use rb.MovePosition, which doesn't actually affect angular velocity. To compromise, I added an if statement that slowly reduces the player's velocity to zero if their steering input is greater than zero (i.e. if they're actively correcting their orientation). I also cleaned up some of the code for reading steering input so more of the calculations are done in FixedUpdate(), to overall reduce the number of necessary calculations.