Spaceship orientation compass and dedicated HUD class
I added code for a rotating compass, to allow the player to easily determine their orientation in the level, since a zero gravity space environment means there’s no ‘up’. This consists of a transform in the spaceship HUD canvas (which has been converted to operate using Screen Space – Camera instead of Screen Space – Overlay), with six cylinders sticking out for each axis in 3D space, and their minus equivalents. Regardless of the player’s current rotation, the transform’s world rotation is always set to zero (plus the relative rotation from the camera to wherever the compass is on the canvas, to ensure the orientation is easier to intuit from the player’s perspective), so it will always show the correct world axes to the player, relative to their current rotation. Each axis is also automatically coloured using what are currently a static set of colour variables, in a new PlanetGrid class (this derives from MonoBehaviour although it currently does not need to be, because I intend on adding more functionality later).
To accomodate these new changes, I created a SpaceshipHUD class. This also includes the speedometer code, which meant I could remove it from the SpaceshipMovement class for reduced clutter.