Overhauled SpaceshipMovement for superior camera control and rotation braking
The previous iteration of my spaceship controller had awkward camera control. The camera would easily lag too far behind the player’s ship, or leave it out of view entirely during sideways movement or rotation.
After cutting up the whole script, and putting it back together, I ended up with a deceptively simple (possibly even crude and primitive) workaround that results in shockingly simple camera movement. Some mechanics have been temporarily removed, most notably the gyro steering, but since that mechanic was not working properly that is fine for now, as I need to rework it and re-implement it later.
Lerping is generally supposed to move between two different stationary values, using a timer that incrementally counts from zero to one. I instead opted to use a hack, by instead specifying the current position, the desired target position, and Time.deltaTime multiplied by a speed variable. The value will partially move from the current position to the desired one in a frame, then do the same in the next, except the updated first position means the distance to travel is smaller. This means the speed of the movement will gradually decelerate as it reaches the target, with shocking smoothness. The catch is that the value will never truly reach the desired target, but for something like a cosmetic camera position, precision is not needed.
I also added an extra mechanic allowing the player to manually brake their rotation by pressing a distinct button (currently Left Alt).
In addition, I also reshuffled some of the variable and function orders, and I removed the warping functionality as I had copied it into a separate class.