Re-implemented more functional spaceship gyro controls
After having to remove the poorly functioning gyro controls in order to re-work the spaceship camera movement, I have now added them back in. They aren’t perfect, but now they work far better than they did before.
I decided to implement the gyro controls into a new class, named GyroSteeringWheel, containing all the specific code to process the gyro values so the player can tilt their phone like a steering wheel or flight joystick. By separating this code from the SpaceshipMovement class, it makes the latter class less complicated and therefore less likely to mess up, and the former code able to be used for any number of scenarios where I want to use this gyro steering code.
This class includes an Update() function that continuously tracks changes to the gyro orientation and therefore the steering value, to ensure that the variable is changed accurately for when it is needed. It also includes a public ResetValues() function that resets the values to zero, in case the player wants to reorient their device.
I also added a new toggle to the spaceship options menu, to enable and disable the gyro controls. This is done by toggling the enabled function of the GyroSteeringWheel on the player controller, as opposed to the toggle directly within the SpaceshipMovement class that I originally planned on using.