Added separate invert functionality to gyro steering
I wanted the X and Y values to be inverted as the 'default' for this steering system, as it made the controls feel more natural and more like using an actual steering wheel or joystick on an aircraft. Rather than awkwardly inverting the values directly in the spaceship movement controller's input, I decided to add customisable invert values into the GyroSteeringWheel class, so it can be easily used for other purposes.