Fixed collision bug with spaceship warping
Turns out, it was caused by a positioning issue. When the new position to move to is calculated, an offset is used to ensure the ship does not crash into the planet it is trying to warp into. I forgot to add the planet’s centre position to the offset value, so the player would warp to the offset relative to 0,0,0 in world space. This meant that the player would often warp inside the sun then have their rigidbody become non-kinematic, causing a collision error that violently shoots the player far outside the level boundaries. By making sure to add the position value, I fixed this error and now the player correctly warps where they need to go.