Updated spaceship steering to be more dependent on velocity, like a conventional plane.
Steer speed is now scaled using the steerSpeedPerVelocityUnit Vector3 value, multiplied by the current velocity magnitude. If the player is stationary, or a bool is enabled to steer independent of velocity, the player will steer at a fixed rotation rate, like before. Additionally, when the ship steers based on velocity, the velocity direction will now update accordingly as well.