Spaceship HUD improvements
I rearranged the HUD elements (save for the pause button since it’s distinct and not part of the normal gameplay) to all occupy the bottom edge of the screen, and also be a bit more symmetrical for improved aesthetics.
I also changed the buttons for braking movement and velocity, and resetting gyro controls, to use new graphics I created to be more informative.