Easier gyro control via holding to activate
Previously, the player gyro controls would always register new rotations, and a button would be present to reset the rotation. This proved to be very awkward, as the player had to constantly turn their phone in awkward directions to compensate for steering, and having to press the reset button repeatedly proved awkward.
I decided to implement a different method – gyro controls are disabled unless the player is holding down a special gyro control button. This is easier to control because the player can easily turn it on and off, and only have it work when they’re trying to turn.
Due to the inevitably subjective nature of which control option would be better, I replaced the ‘enable gyro’ option in the options menu with a toggle to switch between the two modes (having a separate gyro enabled setting would be useless, as the new option functions as gyro controls being disabled if the player never presses the button to activate them).