Increased distance of level boundaries
I noticed that it was possible for the player to warp to planets outside of the map limits and be stuck, with no way to get back except to use the warp function again to another planet that's inside. Although not game-breaking, it was very unprofessional and needed fixing. To this end, I simply moved the level boundary walls further out, so no planet would be stuck outside.