Disabled buttons to access non-functional options menus
If the player is in the third-person spaceship level, the button to access the first-person control options are now disabled. The opposite is also true if the player is in the first-person level. This is because the functionality does not presently exist for altering control options for player 'characters' that the player is not currently using, as there is no way to save data.