Touch inputs are enabled by default when loading a level
This is because I presently don't know a good way to check if a mouse and keyboard is present (with the current input system), but I can check if a touchscreen is. So in order to validate which input system to use depending on the device running the game, it defaults to touch controls and switches off if the system cannot support them.