Planet generator from online database
I have implemented the beginnings of the system for accessing user-generated planets. When the player loads into the spaceship level, instead of randomly generating planets, the game will access a Firebase database to obtain a list of user-generated planets. These planets will then be generated using the data contained in the database. I am very pleased with the core integration, but there are still many features to be implemented.
For example, a large amount of planets in the database will cause hardware strain if too many exist in the level at once, so a mechanic will need to exist where planets are sorted into groups (e.g. solar systems), so the player can choose to travel to (i.e. load in) a specific set of planets at a time.
Additionally, although the code still exists to spawn random planets as a default, there is currently no proper failsafe for if the game cannot establish a proper connection to the online database. I will need to implement a reliable and user-friendly way to inform the player if such an error occurs, so they can either travel to the placeholder planets or take an alternative action.
Finally, obtaining the data and generating each random planet takes time, which will add up in a system dense with user-generated planets. In order to account for this and prevent the user from launching into an empty environment with planets appearing all around them, I will need to implement a loading screen.