Advanced spaceship navigation GUI
I am starting work on a new mechanic that lets the user temporarily show a wide range of additional HUD information, the kind that is generally too obtrusive for normal gameplay. This code required many overhauls to different classes:

SpaceshipNavigationGUI changes:
Added a series of world space compass axes that appear around the sun using LineRenderer classes> This is the main advanced navigation feature that can be turned on and off.
Added code to the SpaceshipNavigationGUI class to show and hide advanced navigation information. The mechanic can be enabled or disabled by tapping the compass on-screen, although there is presently no keyboard command for it.
Added a readonly array of Vector3 values representing the six axis directions. Each set of variables in the class that corresponds to these directions is now an array, such as axis colours and compass points. I made this change so I can use for loops instead of running every axis-related piece of code six times to account for all of them.

Player prefabs:
Added new variables for these new code changes to the Player Spaceship prefab, as well as making the navigation button its own prefab variant for easier standalone editing.
Changed some of the compass axis colours.

SolarSystem class:
This MonoBehaviour class attaches to the GameObject representing the sun, or whatever object is intended to exist at the centre of the level. This contains general information about the current level, that the SpaceshipNavigationGUI class references to display appropriate information.