Loading screen
A current problem with the system for loading user-generated planets is that it takes time to load each planet in. This would result in the player spawning into a level with no planets, and then seeing them appear in real time. This was a problem I needed to fix, and I have done so with a loading screen.
I created the LoadingScreen class, which acts as a MonoBehaviour but has static elements as well. When one of its public static functions, LoadLevel() is called (using a string parameter to specify the desired level to load into). This updates a static string representing the level that needs to open, and loads into a new scene containing a LoadingScreen class and a canvas with GUI elements for information about the level being loaded. An IEnumerator then runs, which loads the new scene additively as an AsyncOperation. Once that is completed, the new level finds and disables all player objects, sets the time scale to zero and unlocks the player cursor. The loading screen does not disappear yet, because the IEnumerator is waiting for additional game processes (such as generating planets) to be completed. Each of these processes registers itself in the loading screen by adding a System.Func<bool> to a static array that the loading IEnumerator checks to ensure that they are all done. Once they are all done, the list is cleared and a button appears. Pressing this button will return the time scale to normal, re-enable the player and unload the loading screen scene.
This was a tricky feature to figure out, due to the complications of getting certain functions to run before control is resumed, but I was able to figure out a deceptively simple solution and get it to work. There are still some bugs I need to iron out, and the prototype loading screen I have isn’t the prettiest, but I am very pleased with how this has all turned out.