Optimised code to force random planet generation
With the old code, the system would go to the process of checking the user-made planet database even if random planet generation was forced, which would mean checking the database was not required. This is now fixed, and if random planets are forced to always generate it will bypass the online code.