Spaceship test level changes – larger solar system, real world planets
I made the play area (and bounding boxes) of the level much larger than before. I also added in several representations of real planets in the solar system, that appear alongside the user-generated planets, on the request of the designer. These new planets can be easily warped to in the same manner as the user-generated or randomly generated planets.