Simple connection test check
Whenever the game needs to check the user’s internet connection (e.g. before loading user-generated planets), the game can consult my new ConnectionTest class. This static class consists of 3 variables – a bool representing the value the last time the status was checked, a float representing the time of the last check, and an IEnumerator for checking. In order to run the proper check, simply run the IEnumerator as a coroutine, then once it’s done, check the bool.
The system is very simple, and perhaps even crude, but works beautifully for the current project.